JavaScriptFilter: Use QScriptEngine to avoid endless loops
QJSEngine does not provide a means to interrupt the
evaluation when then engine stucks in an endless loop.

The (deprecated) QScriptEngine does allow this,
so use this for now. As QBS also needs QScriptEngine,
no extra dependency is created. For Qt installations
without QtScript, the JavaScriptFilter is not compiled.

Task-number: QTCREATORBUG-19690
Change-Id: I172b7e7042eea0d40a765c836b2d1c57b6a529e6
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>