Position environment variables at the top of the environment widget
This ensures that the user doesn't have to scroll down when they've
added a new environment variable and are editing its value; they'll
be able to see its value as they're typing it.

This improves the situation for most cases, but variables whose names
start with letters later in the alphabet cause them fall within the
"end" of the view's range, making it impossible to position them at
the top of the view.

Task-number: QTCREATORBUG-19715
Change-Id: I1f4b5bd311fa2a45770676ba2af3e254ae936f06
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>