Fix ordering of items in open documents popup
When the document history was updated, it was cleaned of any items which
no longer had a document open for them. This worked fine when that only
happened when user closed documents. Nowadays this also happens when
suspending documents without user interaction.

Only remove items from the document history if they are no longer
available even as a suspended document.

Task-number: QTCREATORBUG-19758
Change-Id: I131a24823b5c456879b67a35b039768a15bd7716
Reviewed-by: Konstantin Tokarev <annulen@yandex.ru>
Reviewed-by: David Schulz <david.schulz@qt.io>