KitChooser: Change kit pre-selection strategy
In case there is a startup project, it will be listed in
an additional entry at the top of the kit list, followed
by a separator.

Kits are remembered by Id now, not by index anymore
*unless* it's the startup kit entry, i.e. if the startup
entry is selected and the project is changed, it will
again select the startup entry, possibly indicatin another
kit.

In case a non-startup entry is selected, the kit will be
remembered by id, changing projects will try to re-select
*that* kit.

Change-Id: I827d2670d9a8681570d9c67405cd662cf6a01b4c
Reviewed-by: Orgad Shaneh <orgads@gmail.com>