Squish: Update tst_quick_creation3
Qt 5.3.1 used to be selected automatically which was wrong when the minimum Qt
version is 5.6. Now, because of QTCREATORBUG-19717, this does not happen
anymore.

The kit using Qt 5.6 should be selected automatically, but since we explicitly
do this, the end result is correct. When QTCREATORBUG-19717 is fixed, we
should not need to select a kit ourselves anymore.

Change-Id: I446b2843cd6e545905ce3a6616f7f8d3204bbafe
Reviewed-by: Christian Stenger <christian.stenger@qt.io>