QmlProject: Clean up environment selection
On desktop you get "system environment" as default now and "clean
environment" as option. "kit environment" doesn't make any sense as that
is meant for build configurations.

On remote Devices you only get "clean environment" because we cannot
query the generic remote device for its default environment. However,
as the environment was rather random before and it worked, a clean
environment will probably not break it.

Change-Id: Iab0ed804a21cf77db9e3413926ff710c91561db0
Task-number: QTCREATORBUG-19887
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>