Utils: Watch file replacements with FileSystemWatcher
It's quite common for other programs to remove and re-add a file in
order to change it. QFileSystemWatcher does not track this, but silently
drops the file to be watched.

Add a directory watcher for each directory with files to be watched and
if that triggers, examine if one of the files we are watching in that
directory has been re-added. If so, generate a fileChanged() signal.

This still doesn't track replacements of directories. If we want to
track those, we could apply the same technique to directory watchers,
but we'd have to recursively watch directories down to the file system
root.

Change-Id: I8a70b76bb34502dcfac390ba90e715918d7a5b4a
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Brook Cronin <brook.cronin@qt.io>
Reviewed-by: hjk <hjk@qt.io>