Fix crash in kit's variable chooser if Qt version is removed
Currently the variable chooser cannot update itself if a sub-expander
changes or even is deleted, leading to inconsistencies and crashes if
that happens.
Make the Qt version sub-macroexpander be available for the kit even if
currently is no Qt version selected. For this we create a macro expander
that delegates to the kit's Qt version if possible at variable
substitution time, instead of taking the Qt version's expander directly.

Task-number: QTCREATORBUG-19900
Task-number: QTCREATORBUG-19901
Change-Id: I2bd9b3db60bf6e292abf2a29e0697dc5385709cd
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>