Fix binary editor tool tip
It should show up also over the last column, it should also show up if
the selection spans multiple lines, and if the mouse is not over the
selection it show up with information about the single byte under
cursor. If the mouse is over a selection that is longer than 8 bytes,
the tool tip is restricted to the first 8 bytes of the selection.

Task-number: QTCREATORBUG-17573
Change-Id: Ie7aaf1e52ceadd6be815d4ce3386a60df1418297
Reviewed-by: David Schulz <david.schulz@qt.io>