Remove duplicated "Qt Unit Test" wizard
There are currently two types of wizards for creating auto tests:

- Qt Unit Test
- Auto Test Project

Both produce almost identical code in the case of a Qt GUI-based test,
and the descriptions are very similar. This is really confusing.

The Auto Test Project wizard has support for Qt Test, Google Test and
Qt Quick Test. In addition, the documentation only mentions Auto Test
Projects when it discusses the creation of auto tests:

http://doc.qt.io/qtcreator/creator-autotest.html#creating-tests

So let's just use that and make everything simpler.

Change-Id: Ia16c88c462744baa8df3f9e8082cbcad6ad9eb15
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>