Search Results: Fix expand all button state at startup
It was enabled at startup even though it should only be enabled if
actual search results are shown.
Do not set the enabled state on the button, but on the local action
instead. The button has the Command's action as its default action, and
the state of that action can be overridden by the local action's state
at context change.

Change-Id: I045d022836b12375dc49d4ee0ffd4802d795db3b
Reviewed-by: David Schulz <david.schulz@qt.io>