Enable Clang code model by default
It has been around for a long time now, many e.g. windows specific
issues have been fixed, and we concentrate our development effort on it.
Many bugs in the built-in model stay unfixed, and we'll definitely not
support modern C++ there, which leads to many bug reports that are
closed with "works in clang code model".
Let's switch that around.

Task-number: QTCREATORBUG-19297
Change-Id: I59ec9621ad76590db8ec33f9658af6d103c9da93
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Ivan Donchevskii <ivan.donchevskii@qt.io>