Fix that clearing shortcut in settings did no longer work
This broke in 8d1a592d1637b52e539695afdef7f5564f6d723b

We should ask the validator of a fancy line edit if an entry is valid,
also if the line edit is empty and shows a placeholder text, even if we
then do not mark the place holder with the error color, etc.

In case of the shortcut settings, the validator has side effects that
are important.

Change-Id: I41c0d98ed3c6e9e68df4484a47c108b00ddbbfe7
Reviewed-by: David Schulz <david.schulz@qt.io>