GenericProject: Fix adding several makesteps
Fix adding several makesteps to generic projects again. The problem used
to be that the constructor called by "Step" set a buildTarget of "" via
the delegation to GenericMakeStep's constructor and then added "all"
explicitly. So it was attempting to build two targets, with the empty
target being a syntax error.

Task-number: QTCREATORBUG-20108
Change-Id: I7382a1fa8b8b5869e6be11367e27d6a786f6aa74
Reviewed-by: hjk <hjk@qt.io>