RemoteLinux: Rework RemoteLinuxCustomRunConfiguration
Use ArgumentAspect and WorkingDirectoryAspect.

Use base convenience functions to set up configuration widget.
The RunConfigWidget base is not needed for the main frame.

Make the run config display names consistent.

Display remote exe (which is the main thing) above the local
exe (which is only a helper to get debug info).

Change-Id: I9a47dc9ea27e6abd96bb75a3103ebc8ba35069c0
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>