Core: Avoid duplicating options page
Creating an IOptionPage registers it automatically. In case an
IOptionsProvider is used this duplicates the the option pages
and although duplicated still adds them to the respective
category again.
If one clicks on the respective tab this results in a crash.

Slipped in with cc883023090030eb341a11c4d6634ca027f02c65.

Change-Id: Ie881f532c3f0e397147d99ce368066cdce01a612
Reviewed-by: hjk <hjk@qt.io>