Disable help search widget while indexing documentation
If you try to search while indexing, in the best case you'd get
incomplete results (though with a warning that that is the case), and in
the worst case it crashes because of QTBUG-66816. So just disable the
search widget with info and progress indicator in that case.

Task-number: QTCREATORBUG-20295
Change-Id: Ibec602b1bc0e98f6fef029c6f10c0cdc5197294c
Reviewed-by: Robert Loehning <robert.loehning@qt.io>