Generic/Base make step: Improve details text
The new summary is similar to the one in the qmake make step, adding the
working directory to the summary text, and adding checks for tool chain
and build configuration etc.
Different to the qmake variant is that it doesn't restrict to any
particular build configuration type, and that it needs to show
allArguments() instead of just userArguments().

Change-Id: Ie79264267a2ce834a3d3ffe2e4c066786642adc1
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>