qmake: fix file id mapping lifetime management, take 2
turns out that ref-counting the vfs class is insufficient, as now the
ProFile cache (which expires after a timeout) could outlive all VFS
instances and thus refer to ids which were discarded and later re-used.

to avoid this, we let the cache instance hold a reference to the vfs
class.

amends d03fb3506.

Task-number: QTCREATORBUG-20524
Change-Id: Idd4478ffc1c2405b3b83df32fba45b1f687f6a18
Reviewed-by: Robert Loehning <robert.loehning@qt.io>
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>