Fix that removing splits results in code model issues
The code model is listening to editorsClosed signals, and checks for
which editors are actually visible when that signal is sent.
For this to work correctly, the signal must be sent only at the very end
after deconstructing all the views, because otherwise we might be in a
state where temporarily no editors are "visible".

Fix-up after 80ae992c91985389f6d870ac0201c16d895bc5cc which exposes the
issue.

Task-number: QTCREATORBUG-20464
Change-Id: I9ec5e579d681f2c12924be3d7b89aab398579505
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@qt.io>