Add option for parallel jobs to make step and use it by default
By default the make step uses idealThreadCount jobs, except if it finds
a "-j" option in the MAKEFLAGS variable. The user can opt-in to override
MAKEFLAGS anyhow, and can also choose a different number of build jobs
(including 1).

This option is available for unix-style make. It is not available on
Windows with MSVC tool chain, which would use either nmake (does not
support parallel jobs) or jom (does parallel building by default).

Task-number: QTCREATORBUG-18414
Change-Id: Ic0b5c732d899607ce67ef0b496cc758497a3113a
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>